move yargs logic to launch-from-cli.js
