+++ packages/utils/lib/launch-from-cli.js
@@ -2,6 +2,8 @@ import url from 'url';
 
 import ConfigManager from './config-manager.js';
 import LogManager from './log-manager.js';
+import yargs from 'yargs';
+import { hideBin } from 'yargs/helpers';
 
 /**
  * Resolves with a promise including the current config if your
@@ -31,8 +33,21 @@ export const launchFromCli = async (importMeta, {
 		// eslint-disable-next-line prefer-promise-reject-errors
 		return Promise.reject();
 	}
+
+	let argv = yargs(hideBin(process.argv))
+		.parserConfiguration({
+		// See https://github.com/yargs/yargs-parser#camel-case-expansion
+			'camel-case-expansion': false
+		})
+		.option('config', { alias: 'c', describe: 'Path to your JS or JSON config file.', type: 'string' }).help();
+
+	if (yargsCallback) {
+		argv = yargsCallback(argv);
+	}
+
+	const parsedArgv = await argv.parse();
 	
-	await ConfigManager.getInstance().loadConfig(userConfig, yargsCallback);
+	await ConfigManager.getInstance().loadConfig(userConfig, parsedArgv.config);
 	/** @type {any} TODO: figure out where to add this 'logging' property */
 	const config = ConfigManager.getInstance().getConfig();
 	LogManager.getInstance(config.logging || config);
